Abandoned Colyseus
The program was being built using Colyseus but that was ultimately scrapped as there were struggles setting it up. Instead the new react app uses Socket IO.

Allows users to connect to a page and navigate through a series of join lobby menues. Connections presently work with a few bugs, but the UI does not communicate this well yet